requirements.txt
Django==1.11.20
django-environ==0.4.5
django-extensions==2.1.6
django-filter==2.0.0
djangorestframework==3.9.2
Markdown==3.0.1
pytz==2018.9
six==1.12.0
typing==3.6.6
Unidecode==1.0.23
